feat: /HOC add HOC for talentsList and proofsList
