+++ src/components/TalentsPage/TalentsPage.jsx
@@ -1,42 +1,16 @@
-import { CircularProgress, Grid } from '@mui/material';
-import React, { useEffect } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { Grid } from '@mui/material';
+import React from 'react';
 import { GeneralTalent } from './components/GeneralTalent';
 import { PagesGrid } from '../shared/Grid';
-import { useStoreDispatch } from '../../redux/store';
 import { useSelector } from 'react-redux';
 import {
-	clearList,
-	getGridItem,
 	getGridList,
-	getGridTotalPages,
 	getTalentsList,
-	pendingStatus,
 } from '../../redux/reducers/dataList';
+import { withURL } from '../../HOC/withURL';
 
-export const TalentsPage = () => {
-	const [searchParams, setSearchParams] = useSearchParams();
-
-	const dispatch = useStoreDispatch();
-
+const TalentsPage = ({ total_pages }) => {
 	const talentList = useSelector(getGridList);
-	const total_pages = useSelector(getGridTotalPages);
-	const isLoading = useSelector(pendingStatus);
-	const gridItems = useSelector(getGridItem);
-
-	const urlPage = Number(searchParams.get('page')) || 1;
-
-	useEffect(() => {
-		dispatch(getTalentsList(urlPage - 1));
-
-		return () => dispatch(clearList());
-	}, [urlPage]);
-
-	useEffect(() => {
-		if (urlPage < 0 || (total_pages < urlPage && total_pages !== 0)) {
-			setSearchParams({ page: '1' });
-		}
-	});
 
 	let talentsList = talentList.map(talent => (
 		<Grid item md={6} sm={12} lg={4} key={talent.id}>
@@ -45,13 +19,9 @@ export const TalentsPage = () => {
 	));
 	return (
 		<>
-			{isLoading || gridItems !== 'talents' ? (
-				<div className='loaderContainer'>
-					<CircularProgress />
-				</div>
-			) : (
-				<PagesGrid gridItems={talentsList} total_pages={total_pages} />
-			)}
+			<PagesGrid gridItems={talentsList} total_pages={total_pages} />
 		</>
 	);
 };
+
+export default withURL(TalentsPage, getTalentsList, 'talents');
