+++ src/routes/router.js
@@ -5,8 +5,8 @@ import { Profile } from '../components/Profile';
 import { Provider } from 'react-redux';
 import { store } from '../redux/store';
 import App from '../App';
-import { TalentsPage } from '../components/TalentsPage/TalentsPage';
-import { ProofPage } from '../components/ProofPage/ProofPage';
+import  TalentsPage  from '../components/TalentsPage/TalentsPage';
+import  ProofPage  from '../components/ProofPage/ProofPage';
 import { CreateProof } from '../components/CreateProof';
 export const router = createBrowserRouter([
 	{
